Move restaurant and address models
